# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.LeaveListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.LeaveRequest'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
  dto.PayrollListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Payroll'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_page:
        type: integer
    type: object
...
  models.LeaveRequest:
    properties:
      employee_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      leave_type:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
  models.Payroll:
    properties:
      base_salary:
        type: number
      bonus:
        type: number
      deductions:
        type: number
      employee_id:
        type: integer
      id:
        type: integer
      month:
        description: could be first day of month
        type: string
      net_salary:
        type: number
    type: object
...
info:
  contact:
    email: support@hrms.example.com
    name: HRMS Support
    url: http://hrms.example.com/support
  description: Human Resource Management System API built with Go (Gin framework),
    GORM, and MySQL. Supports Employee management, Attendance, Payroll, Leave Requests,
    Departments, Job Titles, and User Authentication/Authorization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://hrms.example.com/terms/
  title: HRMS API
  version: "1.0"
paths:
...
  /leaves:
    get:
      description: Get paginated list of leave requests
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LeaveListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List leave requests with pagination
      tags:
      - Leaves
    post:
      consumes:
      - application/json
      description: Create a new leave request for an employee
      parameters:
      - description: Leave Request object
        in: body
        name: leave
        required: true
        schema:
          $ref: '#/definitions/models.LeaveRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LeaveRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new leave request
      tags:
      - Leaves
  /leaves/{id}:
    delete:
      description: Delete a leave request by ID
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a leave request
      tags:
      - Leaves
    get:
      description: Retrieve leave request details by ID
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LeaveRequest'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get leave request by ID
      tags:
      - Leaves
    put:
      consumes:
      - application/json
      description: Update leave request details by ID
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Leave Request object
        in: body
        name: leave
        required: true
        schema:
          $ref: '#/definitions/models.LeaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LeaveRequest'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a leave request
      tags:
      - Leaves
  /leaves/{id}/approve:
    patch:
      description: Approve leave request by ID
      parameters:
      - description: Leave Request ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Approve a leave request
      tags:
      - Leaves
...
  /payrolls:
    get:
      consumes:
      - application/json
      description: Get paginated list of payrolls, optionally filter by employee_id
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by employee ID
        in: query
        name: employee_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PayrollListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List Payrolls
      tags:
      - Payroll
    post:
      consumes:
      - application/json
      description: Create a new payroll record for an employee
      parameters:
      - description: Payroll data
        in: body
        name: payroll
        required: true
        schema:
          $ref: '#/definitions/models.Payroll'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Payroll'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create Payroll
      tags:
      - Payroll
  /payrolls/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a payroll record by ID
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete Payroll
      tags:
      - Payroll
    get:
      consumes:
      - application/json
      description: Get a payroll record by ID
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payroll'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get Payroll
      tags:
      - Payroll
    put:
      consumes:
      - application/json
      description: Update a payroll record by ID
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated payroll data
        in: body
        name: payroll
        required: true
        schema:
          $ref: '#/definitions/models.Payroll'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Payroll'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update Payroll
      tags:
      - Payroll
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by your JWT token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"